pyproject.toml
[project]
name = "simplefastapiapp"
version = "0.1.0"
description = ""
authors = [
    {name = "Fadel Bidiga",email = "bidigafadel@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "annotated-types (==0.7.0)",
    "anyio (==4.9.0)",
    "asgiref (==3.9.1)",
    "certifi (==2025.7.14)",
    "charset-normalizer (==3.4.2)",
    "click (==8.2.1)",
    "fastapi (==0.115.14)",
    "googleapis-common-protos (==1.70.0)",
    "grpcio (==1.73.1)",
    "h11 (==0.16.0)",
    "idna (==3.10)",
    "importlib_metadata (==8.7.0)",
    "iniconfig (==2.1.0)",
    "opentelemetry-api (==1.35.0)",
    "opentelemetry-exporter-otlp (==1.35.0)",
    "opentelemetry-exporter-otlp-proto-common (==1.35.0)",
    "opentelemetry-exporter-otlp-proto-grpc (==1.35.0)",
    "opentelemetry-exporter-otlp-proto-http (==1.35.0)",
    "opentelemetry-instrumentation (==0.56b0)",
    "opentelemetry-instrumentation-asgi (==0.56b0)",
    "opentelemetry-instrumentation-fastapi (==0.56b0)",
    "opentelemetry-proto (==1.35.0)",
    "opentelemetry-sdk (==1.35.0)",
    "opentelemetry-semantic-conventions (==0.56b0)",
    "opentelemetry-util-http (==0.56b0)",
    "packaging (==25.0)",
    "pluggy (==1.6.0)",
    "protobuf (==6.31.1)",
    "pydantic (==2.11.7)",
    "pydantic-settings (==2.10.1)",
    "pydantic-core (==2.33.2)",
    "pygments (==2.19.2)",
    "python-dotenv (==1.1.1)",
    "requests (==2.32.4)",
    "sniffio (==1.3.1)",
    "sqlalchemy (==2.0.41)",
    "starlette (==0.46.2)",
    "typing-inspection (==0.4.1)",
    "typing-extensions (==4.14.1)",
    "urllib3 (==2.5.0)",
    "uvicorn (==0.35.0)",
    "wrapt (==1.17.2)",
    "zipp (==3.23.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "redis (>=6.2.0,<7.0.0)",
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
